FourDaubScaling more robust
Make sure the necessary number of factors are included in the product
defining the scaling function.
An upper limit on the number of factors is still needed: When (e.g.) the
input is 0, the low pass filter does not evaluate to 1 and the while
loop does not terminate with prec=eps().